pyproject.toml
[project]
name = "taskio"
version = "0.1.0"
description = "None"
readme = "README.md"
authors = [{ name = "pro2015", email = "" }]
requires-python = ">=3.12"
dependencies = [
    "fletxr==0.1.4.a0",
    "flet[all]",
    "peewee>=3.18.1",
    "newsapi-python>=0.2.7",
    "python-dotenv>=1.1.1",
    "newsdataapi>=0.1.25",
]
